+++ CHANGELOG.md
@@ -5,6 +5,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [5.9.0] - 2024-07-29
+
 ### Added
 
 - upcloud_managed_database_\*: support for labels
@@ -689,7 +691,8 @@ Updated upcloud-go-api, added build/CI scripts, and repackaged 0.1.0 as 1.0.0.
 - resource_upcloud_firewall_rule removed and replaced by resource_upcloud_firewall_rules
 - resource_upcloud_zone removed and replaced by zone and zones datasources
 
-[Unreleased]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.8.1...HEAD
+[Unreleased]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.9.0...HEAD
+[5.9.0]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.8.1...v5.9.0
 [5.8.1]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.8.0...v5.8.1
 [5.8.0]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.7.0...v5.8.0
 [5.7.0]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.6.1...v5.7.0